requirements.txt
Flask
Pillow
onnxruntime
numpy
requests
torch
torchvision
urllib3
